Refactor OnLook to seperate controller, fix bug where player could walk while not grounded and add drag
