+++ AoTRetaliationOfTheSurveyCorps/Assets/Camera/FollowTarget.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class CameraController : MonoBehaviour
+public class FollowTarget: MonoBehaviour
 {
     [Header("References")]
     [SerializeField]
