+++ AoTRetaliationOfTheSurveyCorps/Assets/Characters/Player/LookController.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 6e0629edfb044fe5b4c6100ea35c73d7
+timeCreated: 1680112259
