+++ AoTRetaliationOfTheSurveyCorps/Assets/OdmGear/Gas.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a17c3a20ae7707540902aa4835cff8c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
