+++ AoTRetaliationOfTheSurveyCorps/Assets/OdmGear/GrappleHooks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a4b6acd2d83d7e64c850c0025490e577
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
